pyproject.toml
[tool.poetry]
name = "board_members_information_extractor"
version = "0.0.1"
description = ""
authors = ["mzhadigerov <mzhadigerov@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.6.2"
tika = "^1.24"
bs4 = "^0.0.1"
camelot-py = {extras = ["base"], version = "^0.10.1"}
podder-task-foundation = {git = "https://github.com/podder-ai/podder-task-foundation.git", rev = "main"}
spacy = "^3.1.2"
spacy-transformers = "^1.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.6.0"
yapf = "^0.30.0"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"